refactor: extract duplicate grammar loading logic to GrammarLoader
Create a shared GrammarLoader utility enum that provides synchronous
grammar loading from Homebrew and cache directories. Both
TreeSitterHighlighter and LazyTreeSitterHighlighter now delegate to
this shared implementation instead of having duplicate code.

This removes ~50 lines of duplicate code from each highlighter and
ensures both stay in sync for grammar loading behavior.

Closes #48
